+++ CHANGELOG.rst
@@ -9,6 +9,9 @@ The format is based on `Keep a Changelog`_, and this project adheres to `Semanti
 `Unreleased`_
 -------------
 
+`4.1.0`_ - 2022-10-28
+---------------------
+
 Added
 ^^^^^
 * Allow assuming phony dependencies as up-to-date
@@ -387,8 +390,9 @@ Added
 * Initial project structure
 
 
-.. _`unreleased`: https://github.com/spapanik/yamk/compare/v4.0.0...main
-.. _`4.0.4`: https://github.com/spapanik/yamk/compare/v3.0.1...v4.0.0
+.. _`unreleased`: https://github.com/spapanik/yamk/compare/v4.1.0...main
+.. _`4.1.0`: https://github.com/spapanik/yamk/compare/v4.0.0...v4.1.0
+.. _`4.0.0`: https://github.com/spapanik/yamk/compare/v3.0.1...v4.0.0
 .. _`3.0.1`: https://github.com/spapanik/yamk/compare/v3.0.0...v3.0.1
 .. _`3.0.0`: https://github.com/spapanik/yamk/compare/v2.4.1...v3.0.0
 .. _`2.4.1`: https://github.com/spapanik/yamk/compare/v2.4.0...v2.4.1